+++ dartsass/cli/compile.py
@@ -1,0 +1,19 @@
+import logging
+
+import click
+
+from ..compiler import DartSassCompiler
+from ..exceptions import RunnedCommandError
+
+
+@click.command()
+@click.pass_context
+def compile_command(context):
+    """
+    R&D on managing subprocess success and failure from command execution
+    """
+    logger = logging.getLogger("flechette-insolente")
+
+    compiler = DartSassCompiler()
+
+    logger.info("Foo.")