+++ dartsass/cli/entrypoint.py
@@ -3,11 +3,12 @@ Main entrance to commandline actions
 """
 import click
 
-from dartsass.logger import init_logger
+from ..logger import init_logger
 
-from dartsass.cli.version import version_command
-from dartsass.cli.greet import greet_command
-
+from .version import version_command
+from .greet import greet_command
+from .compile import compile_command
+from .exec_dev import execdev_command
 
 # Help alias on "-h" argument
 CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
@@ -59,3 +60,5 @@ def cli_frontend(ctx, verbose):
 # Attach commands methods to the main grouper
 cli_frontend.add_command(version_command, name="version")
 cli_frontend.add_command(greet_command, name="greet")
+cli_frontend.add_command(compile_command, name="compile")
+cli_frontend.add_command(execdev_command, name="execdev")