+++ dartsass/sass_version.py
@@ -1,4 +1,0 @@
-from dartsass.exceptions import AppOperationError
-from dartsass.embedded_sass_pb2 import InboundMessage, OutboundMessage
-
-print("Hello world!")