+++ docs/history.rst
@@ -4,7 +4,7 @@
 History
 =======
 
-Version 0.3.0 - Unreleased
+Version 0.3.0 - 2023/10/04
 **************************
 
 Totally changed spirit of this package, we dropped the idea of using *Embedded Host*
@@ -13,6 +13,13 @@ communication, etc..).
 
 Instead we move this package as a Python wrapper that will use subprocess to execute
 dart-sass binary.
+
+At this stage, the command line ``compile`` is working correctly but it stills an
+alpha prototype since it only support Linux 64bits plateform until we properly. Also
+for now only a few set of dart-sass parameters are implemented.
+
+This release won't be released on Pypi since we should implement the right way to get
+sub packages for all executable plateforms.
 
 
 Version 0.2.0 - Unreleased